+++ src/phenoml/construe/__init__.py
@@ -12,6 +12,9 @@ if typing.TYPE_CHECKING:
         CodeResponse,
         CodeSystemDetails,
         CodeSystemInfo,
+        CrosswalkMatch,
+        CrosswalkResponse,
+        CrosswalkTarget,
         DeleteCodeSystemResponse,
         ExportCodeSystemResponse,
         ExtractCodesResult,
@@ -34,8 +37,10 @@ if typing.TYPE_CHECKING:
         TextSearchResult,
     )
     from .errors import (
+        BadGatewayError,
         BadRequestError,
         ConflictError,
+        ContentTooLargeError,
         FailedDependencyError,
         ForbiddenError,
         GatewayTimeoutError,
@@ -48,6 +53,7 @@ if typing.TYPE_CHECKING:
     from . import code_systems, codes
     from .code_systems import UploadRequestFormat, UploadResponse
 _dynamic_imports: typing.Dict[str, str] = {
+    "BadGatewayError": ".errors",
     "BadRequestError": ".errors",
     "Citation": ".types",
     "CodeCategory": ".types",
@@ -55,6 +61,10 @@ _dynamic_imports: typing.Dict[str, str] = {
     "CodeSystemDetails": ".types",
     "CodeSystemInfo": ".types",
     "ConflictError": ".errors",
+    "ContentTooLargeError": ".errors",
+    "CrosswalkMatch": ".types",
+    "CrosswalkResponse": ".types",
+    "CrosswalkTarget": ".types",
     "DeleteCodeSystemResponse": ".types",
     "ExportCodeSystemResponse": ".types",
     "ExtractCodesResult": ".types",
@@ -112,6 +122,7 @@ def __dir__():
 
 
 __all__ = [
+    "BadGatewayError",
     "BadRequestError",
     "Citation",
     "CodeCategory",
@@ -119,6 +130,10 @@ __all__ = [
     "CodeSystemDetails",
     "CodeSystemInfo",
     "ConflictError",
+    "ContentTooLargeError",
+    "CrosswalkMatch",
+    "CrosswalkResponse",
+    "CrosswalkTarget",
     "DeleteCodeSystemResponse",
     "ExportCodeSystemResponse",
     "ExtractCodesResult",